Fix compile errors in `assistant_slash_commands`
